lp ticker - fix until methods - time is in period addition
We should check if now and until are valid compare to the current time.
It can happen that real_now is far in the future, we should not sleep or set interrupt,
just return.

K64F is using deltas for setting the interrupt, we should use real_now
delta = time - real_now